Smarter flow:
Trying to condense the process into a loop. My idea now is to have a JSON (or txt, or other) file with a name/URL key pair (i.e. "Theme Name":"https://themeurl.com"), which the loop will run over to create a folder in the themes/fonts/etc folder with the "name" portion of they key pair, and then clone/download the theme into that folder.